dual-source-webscraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
pandas==2.2.2
matplotlib==3.9.0
playwright==1.48.0
tqdm==4.66.4
pyyaml==6.0.2
tenacity==8.3.0
streamlit==1.37.1
pytest==8.2.1
python-dateutil==2.9.0.post0
